requirements.txt
python-dotenv
boto3
opencv-python
pandas 
numpy 
matplotlib
torch
albumentations
segmentation-models-pytorch
matplotlib
opencv-python
Pillow
tqdm
scikit-learn

dvc
notebook
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy
Flask
Flask-Cors
mlflow